+++ Source/OxyPlot/Rendering/TextVerticalAlignment.cs
@@ -1,0 +1,37 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="TextVerticalAlignment.cs" company="OxyPlot">
+//   Copyright (c) 2020 OxyPlot contributors
+// </copyright>
+// <summary>
+//   Specifies vertical alignment for text.
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace OxyPlot
+{
+    /// <summary>
+    /// Specifies vertical alignment for text.
+    /// </summary>
+    public enum TextVerticalAlignment
+    {
+        /// <summary>
+        /// Aligned at the top.
+        /// </summary>
+        Top = -1,
+
+        /// <summary>
+        /// Aligned in the middle.
+        /// </summary>
+        Middle = 0,
+
+        /// <summary>
+        /// Aligned at the bottom.
+        /// </summary>
+        Bottom = 1,
+
+        /// <summary>
+        /// Aligned at the baseline.
+        /// </summary>
+        Baseline = 2,
+    }
+}